cast int to long for older revit versions

+++ PowerBuilder/Commands/pcmdUpdateViewTemplateByTags.cs
@@ -143,11 +143,11 @@ namespace PowerBuilder.Commands
                 //redundant category checks should be eliminated
                 //if we're doing a 'transparent' overrides merge, then redundant category checking is probably ok.
                 foreach (ElementId lap in LayerAffectedParameters) {
-                    int lapValue;
+                    long lapValue;
 #if REVIT2024_OR_GREATER
                     lapValue = lap.Value;
 #else
-                    lapValue = lap.IntegerValue;
+                    lapValue = (long)lap.IntegerValue;
 #endif
                     if (_overrideParameters.Contains((BuiltInParameter)(lapValue))) {
                         MergeGraphicOverrideStrategy((BuiltInParameter)(lapValue), ViewTemplate, ViewLayer);
